Make Graph generation faster

+++ benchmark/inc/graph.hpp
@@ -5,8 +5,8 @@
 
 // Edge Struct
 struct Edge {
-    size_t outgoing;
-    size_t weight;
+    uint16_t outgoing;
+    uint16_t weight;
 };
 
 // Abstract base class for graphs
@@ -14,15 +14,15 @@ class Graph {
   private:
 
   protected:
-    const size_t v_;
+    const uint16_t v_;
 
   public:
-      Graph(size_t v);
+      Graph(uint16_t v);
       virtual ~Graph() = default;
 
-      virtual void addEdge(size_t start, size_t end, size_t weight) = 0;
-      virtual std::vector<Edge> edgesFromStart(size_t start) const = 0;
-      size_t getNumVertices() const;
+      virtual void addEdge(uint16_t start, uint16_t end, uint16_t weight) = 0;
+      virtual std::vector<Edge> edgesFromStart(uint16_t start) const = 0;
+      uint16_t getNumVertices() const;
 };
 
 
@@ -32,21 +32,21 @@ private:
     std::vector<std::vector<Edge>> adjList_;
 
 public:
-    GraphAdjList(size_t v);
+    GraphAdjList(uint16_t v);
     ~GraphAdjList();
 
-    void addEdge(size_t start, size_t end, size_t weight) override;
-    std::vector<Edge> edgesFromStart(size_t start) const override;
+    void addEdge(uint16_t start, uint16_t end, uint16_t weight) override;
+    std::vector<Edge> edgesFromStart(uint16_t start) const override;
 };
 
 class GraphAdjMatrix : public Graph {
-    std::vector<size_t> adjMatrix_;
+    std::vector<uint16_t> adjMatrix_;
 
     public:
-        GraphAdjMatrix(size_t v);
+        GraphAdjMatrix(uint16_t v);
         ~GraphAdjMatrix();
 
-    void addEdge(size_t start, size_t end, size_t weight) override;
-    std::vector<Edge> edgesFromStart(size_t start) const override;
+    void addEdge(uint16_t start, uint16_t end, uint16_t weight) override;
+    std::vector<Edge> edgesFromStart(uint16_t start) const override;
 
 };
